pyproject.toml
[tool.poetry]
name = "fastapi-service-template"
version = "0.1.0"
description = ""
authors = ["Duc-Minh Phan <alephvn@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = { version = "^2.0.19", extras = ["asyncio"] }
aiomysql = "^0.1.1"
alembic = "^1.11.1"
fastapi = "^0.100.0"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.7"
httpx = "^0.24.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
asgiref = "^3.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "F", # Pyflakes
    "E", # Pycodestyle
    "W", # Pycodestyle
    "I", # isort
    "S", # flake8-bandit
    "COM", # flake8-commas
    "ISC", # flake8-implicit-str-concat
    "T20", # flake8-print
    "UP", # pyupgrade
    "RUF", # Ruff-specific rules
]

ignore = [
    "S101",
    "S105",
    "S301",
    # need to be fixed for Pydantic models:
    # https://github.com/astral-sh/ruff/issues/5243
    "RUF012",
]

extend-exclude = ["migrations"]

# Same as Black.
line-length = 88

target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
follow_imports = "silent"
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["main"]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
]
